requirements.txt
flask>=3.0
python-dotenv>=1.0
requests>=2.31
Pillow>=10.0
gunicorn>=21.2
